updated route functionality

+++ router/blog_router.py
@@ -1,33 +1,21 @@
 from fastapi import APIRouter, Depends, HTTPException, status
 from service.Blog_services import reusables_codes
 from sqlalchemy.orm import Session
-from schema.article_schema import ShowBlog, BlogCreate
+from schema.article_schema import ShowBlog, BlogCreate, Blog
 from schema.models import Blogs
 from datetime import datetime
-from typing import List
 from router.login_router import oauth2_scheme
 
 
 blog_route = APIRouter()
 
-#FILTER ARTICLES BY AUTHOR ID
-@blog_route.get("/{user_id}", response_model= List[ShowBlog])
-async def list_blog_from_author(user_id, db:Session = Depends(reusables_codes.get_db)):
-    articles = db.query(Blogs).filter(Blogs.owner_id == user_id)
-    if not articles.first():
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail='No articles from this user')
-    
-    return articles
-
 #CREATE A NEW ARTICLE
 @blog_route.post("/create", response_model= ShowBlog)
-async def create_article(article:BlogCreate, db:Session = Depends(reusables_codes.get_db), token:str=Depends(oauth2_scheme)):
+async def create_article(title: str, article:BlogCreate, db:Session = Depends(reusables_codes.get_db), token:str=Depends(oauth2_scheme)):
     
-    user = reusables_codes.get_user_from_token(db, token)           #authorization
+    user = reusables_codes.get_user_from_token(db, token)                                       #authorization
     
-    new_article = Blogs(**article.dict(), date_posted = datetime.now().date(), owner_id = user.id, author = user.firstname+' '+user.lastname)
+    new_article = Blogs(title = title, **article.dict(), date_posted = datetime.now().date(), owner_id = user.id, author = user.firstname+' '+user.lastname)
     db.add(new_article)
     db.commit()
     db.refresh(new_article)
@@ -35,10 +23,25 @@ async def create_article(article:BlogCreate, db:Session = Depends(reusables_code
             status_code=status.HTTP_201_CREATED, 
             detail='article created successfully',
         )
+    
+    
+#FILTER ARTICLES BY AUTHOR ID
+@blog_route.get("/view_blogs", response_model= list[ShowBlog])
+async def view_my_blog_posts(db:Session = Depends(reusables_codes.get_db), token:str=Depends(oauth2_scheme)):
+    
+    user = reusables_codes.get_user_from_token(db, token)           #authorization
+
+    articles = db.query(Blogs).filter(Blogs.owner_id == user.id)
+    if not articles.first():
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail='No articles from this user')
+    
+    return articles
 
 #EDITING AN ARTICLE BY Author ONLY
 @blog_route.put("/update/{id}")
-async def Edit_article(id:int, input:BlogCreate, db:Session = Depends(reusables_codes.get_db), token:str=Depends(oauth2_scheme)):
+async def Edit_my_article(id:int, input:Blog, db:Session = Depends(reusables_codes.get_db), token:str=Depends(oauth2_scheme)):
     
     user = reusables_codes.get_user_from_token(db, token)           #authorization
     
@@ -67,7 +70,7 @@ async def Edit_article(id:int, input:BlogCreate, db:Session = Depends(reusables_
 
 #DELETE AN ARTICLE BY Author ONLY
 @blog_route.delete("/delete/{id}")
-async def delete_article(id:int, db:Session = Depends(reusables_codes.get_db), token:str=Depends(oauth2_scheme)):
+async def delete_article(id:int, password:str, db:Session = Depends(reusables_codes.get_db), token:str=Depends(oauth2_scheme)):
     
     user = reusables_codes.get_user_from_token(db, token)      #authorization
     
@@ -77,7 +80,7 @@ async def delete_article(id:int, db:Session = Depends(reusables_codes.get_db), t
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Article id:{id} does not exist")
     
-    if existing_article.first().owner_id == user.id:
+    if existing_article.first().owner_id == user.id and user.password == password:
         existing_article.delete() 
         db.commit()
         raise HTTPException(
@@ -87,5 +90,5 @@ async def delete_article(id:int, db:Session = Depends(reusables_codes.get_db), t
         
     raise HTTPException(
             status_code=status.HTTP_403_FORBIDDEN, 
-            detail= 'An article can only be DELETED by its Authors'
+            detail= 'This article can only be DELETED by authenticated Authors'
         )
